+++ web-frontend/modules/database/components/row/RowEditModal.vue
@@ -11,7 +11,7 @@
       :field="field"
       :row="row"
       @update="update"
-      @field-updated="$emit('field-updated')"
+      @field-updated="$emit('field-updated', $event)"
       @field-deleted="$emit('field-deleted')"
     ></RowEditModalField>
     <div class="actions">